+++ app/src/main/kotlin/com/wire/android/navigation/NavigationGraph.kt
@@ -22,18 +22,14 @@ package com.wire.android.navigation
 
 import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
 import com.google.accompanist.navigation.material.ExperimentalMaterialNavigationApi
 import com.ramcosta.composedestinations.DestinationsNavHost
 import com.ramcosta.composedestinations.animations.defaults.NestedNavGraphDefaultAnimations
 import com.ramcosta.composedestinations.animations.defaults.RootNavGraphDefaultAnimations
 import com.ramcosta.composedestinations.animations.rememberAnimatedNavHostEngine
-import com.ramcosta.composedestinations.navigation.dependency
 import com.ramcosta.composedestinations.spec.Route
 import com.wire.android.ui.NavGraphs
-import com.wire.android.ui.authentication.create.common.CreateAccountViewModel
 
 @OptIn(ExperimentalMaterialNavigationApi::class, ExperimentalAnimationApi::class)
 @Composable
@@ -43,10 +39,11 @@ fun NavigationGraph(
 ) {
     val navHostEngine = rememberAnimatedNavHostEngine(
         rootDefaultAnimations = RootNavGraphDefaultAnimations.ACCOMPANIST_FADING,
-        defaultAnimationsForNestedNavGraph = mapOf(NavGraphs.createPersonalAccount to NestedNavGraphDefaultAnimations(
-            enterTransition = { smoothSlideInFromRight() },
-            exitTransition = { smoothSlideOutFromLeft() }
-        ),
+        defaultAnimationsForNestedNavGraph = mapOf(
+            NavGraphs.createPersonalAccount to NestedNavGraphDefaultAnimations(
+                enterTransition = { smoothSlideInFromRight() },
+                exitTransition = { smoothSlideOutFromLeft() }
+            ),
             NavGraphs.createTeamAccount to NestedNavGraphDefaultAnimations(
                 enterTransition = { smoothSlideInFromRight() },
                 exitTransition = { smoothSlideOutFromLeft() }
@@ -55,20 +52,6 @@ fun NavigationGraph(
     )
 
     DestinationsNavHost(
-        dependenciesContainerBuilder = {
-            dependency(NavGraphs.createPersonalAccount) {
-                val parentEntry = remember(navBackStackEntry) {
-                    navController.getBackStackEntry(NavGraphs.createPersonalAccount.route)
-                }
-                hiltViewModel<CreateAccountViewModel>(parentEntry)
-            }
-            dependency(NavGraphs.createTeamAccount) {
-                val parentEntry = remember(navBackStackEntry) {
-                    navController.getBackStackEntry(NavGraphs.createTeamAccount.route)
-                }
-                hiltViewModel<CreateAccountViewModel>(parentEntry)
-            }
-        },
         navGraph = NavGraphs.root,
         engine = navHostEngine,
         startRoute = startDestination,
