+++ app/src/main/kotlin/com/wire/android/ui/authentication/create/email/CreateAccountEmailScreen.kt
@@ -65,7 +65,6 @@ import com.wire.android.R
 import com.wire.android.ui.authentication.ServerTitle
 import com.wire.android.ui.authentication.create.common.CreateAccountFlowType
 import com.wire.android.ui.authentication.create.common.CreateAccountNavArgs
-import com.wire.android.ui.authentication.create.common.CreateAccountViewModel
 import com.wire.android.ui.authentication.create.common.CreatePersonalAccountNavGraph
 import com.wire.android.ui.authentication.create.common.CreateTeamAccountNavGraph
 import com.wire.android.ui.common.WireDialog
@@ -93,7 +92,6 @@ import com.wire.kalium.logic.configuration.server.ServerConfig
 @Destination(navArgsDelegate = CreateAccountNavArgs::class)
 @Composable
 fun CreateAccountEmailScreen(
-    viewModel: CreateAccountViewModel,
     createAccountEmailViewModel: CreateAccountEmailViewModel = hiltViewModel(),
     navigator: DestinationsNavigator
 ) {
@@ -116,7 +114,7 @@ fun CreateAccountEmailScreen(
             onTermsAccept = ::onTermsAccept,
             onErrorDismiss = ::onEmailErrorDismiss,
             tosUrl = tosUrl(),
-            serverConfig = viewModel.serverConfig
+            serverConfig = serverConfig
         )
     }
 }
